refactor: move datagrid icon and styles to separate files
